// client-frontend/src/shared/types/query-keys.ts
export type QueryKeys =
  | "assigned-rates"
  | "self-rates"
  | "confirm-rates"
  | "confirm-curator-rates"
  | "support"
  | "users"
  | "teams"
  | "rates-me"
  | "ipr-me"
  | "user-board"
  | "report";

export const queryKeys = {
  assignedRates: "assigned-rates",
  selfRates: "self-rates",
  confirmRates: "confirm-rates",
  confirmCuratorRates: "confirm-curator-rates",

  support: "support",
  users: "users",
  teams: "teams",
  ratesMe: "rates-me",

  iprMe: "ipr-me",
  userBoard: "user-board",

  report: "report",
};